+++ injectproxy/routes.go
@@ -25,14 +25,15 @@ import (
 )
 
 type routes struct {
-	upstream  *url.URL
-	handler   http.Handler
-	label     string
+	upstream *url.URL
+	handler  http.Handler
+	label    string
+
 	mux       *http.ServeMux
 	modifiers map[string]func(*http.Response) error
 }
 
-func NewRoutes(upstream *url.URL, label string) *routes {
+func NewRoutes(upstream *url.URL, label string, nonAPIPathPassthrough bool) *routes {
 	proxy := httputil.NewSingleHostReverseProxy(upstream)
 
 	r := &routes{
@@ -41,6 +42,10 @@ func NewRoutes(upstream *url.URL, label string) *routes {
 		label:    label,
 	}
 	mux := http.NewServeMux()
+
+	if nonAPIPathPassthrough {
+		mux.Handle("/", http.HandlerFunc(r.noop))
+	}
 	mux.Handle("/federate", enforceMethods(r.federate, "GET"))
 	mux.Handle("/api/v1/query", enforceMethods(r.query, "GET", "POST"))
 	mux.Handle("/api/v1/query_range", enforceMethods(r.query, "GET", "POST"))
@@ -49,6 +54,7 @@ func NewRoutes(upstream *url.URL, label string) *routes {
 	mux.Handle("/api/v2/silences", enforceMethods(r.silences, "GET", "POST"))
 	mux.Handle("/api/v2/silences/", enforceMethods(r.silences, "GET", "POST"))
 	mux.Handle("/api/v2/silence/", enforceMethods(r.deleteSilence, "DELETE"))
+	mux.Handle("/api/", http.HandlerFunc(r.notImplemented))
 	r.mux = mux
 	r.modifiers = map[string]func(*http.Response) error{
 		"/api/v1/rules":  modifyAPIResponse(r.filterRules),
@@ -65,6 +71,7 @@ func (r *routes) ServeHTTP(w http.ResponseWriter, req *http.Request) {
 		return
 	}
 	req = req.WithContext(withLabelValue(req.Context(), lvalue))
+
 	// Remove the proxy label from the query parameters.
 	q := req.URL.Query()
 	q.Del(r.label)
@@ -115,6 +122,10 @@ func withLabelValue(ctx context.Context, label string) context.Context {
 
 func (r *routes) noop(w http.ResponseWriter, req *http.Request) {
 	r.handler.ServeHTTP(w, req)
+}
+
+func (r *routes) notImplemented(w http.ResponseWriter, _ *http.Request) {
+	http.Error(w, "Not implemented", http.StatusNotImplemented)
 }
 
 func (r *routes) query(w http.ResponseWriter, req *http.Request) {