+++ injectproxy/routes_test.go
@@ -97,12 +97,12 @@ func (m *mockUpstream) Close() {
 
 const proxyLabel = "namespace"
 
-func TestEndpointNotImplemented(t *testing.T) {
+func TestNonAPIEndpointNotFound(t *testing.T) {
 	m := newMockUpstream(http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
 		w.Write(okResponse)
 	}))
 	defer m.Close()
-	r := NewRoutes(m.url, proxyLabel)
+	r := NewRoutes(m.url, proxyLabel, false)
 
 	w := httptest.NewRecorder()
 	req := httptest.NewRequest("GET", "http://prometheus.example.com/graph?namespace=ns1", nil)
@@ -111,6 +111,44 @@ func TestEndpointNotImplemented(t *testing.T) {
 	resp := w.Result()
 	if resp.StatusCode != http.StatusNotFound {
 		t.Fatalf("expected status code 404, got %d", resp.StatusCode)
+	}
+}
+
+func TestNonAPIEndpointPassthrough(t *testing.T) {
+	m := newMockUpstream(http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
+		w.Write(okResponse)
+	}))
+	defer m.Close()
+	r := NewRoutes(m.url, proxyLabel, true)
+
+	w := httptest.NewRecorder()
+	req := httptest.NewRequest("GET", "http://prometheus.example.com/graph?namespace=ns1", nil)
+
+	r.ServeHTTP(w, req)
+	resp := w.Result()
+	if resp.StatusCode != http.StatusOK {
+		t.Fatalf("expected status code 200, got %d", resp.StatusCode)
+	}
+}
+
+func TestAPIEndpointNotImplemented(t *testing.T) {
+	m := newMockUpstream(http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
+		w.Write(okResponse)
+	}))
+	defer m.Close()
+	for _, nonAPIPassthrough := range []bool{true, false} {
+		t.Run(fmt.Sprintf("passthrough=%v", nonAPIPassthrough), func(t *testing.T) {
+			r := NewRoutes(m.url, proxyLabel, nonAPIPassthrough)
+
+			w := httptest.NewRecorder()
+			req := httptest.NewRequest("GET", "http://prometheus.example.com/api/unknown?namespace=ns1", nil)
+
+			r.ServeHTTP(w, req)
+			resp := w.Result()
+			if resp.StatusCode != http.StatusNotImplemented {
+				t.Fatalf("expected status code 501, got %d", resp.StatusCode)
+			}
+		})
 	}
 }
 
@@ -151,7 +189,7 @@ func TestFederate(t *testing.T) {
 				),
 			)
 			defer m.Close()
-			r := NewRoutes(m.url, proxyLabel)
+			r := NewRoutes(m.url, proxyLabel, false)
 
 			u, err := url.Parse("http://prometheus.example.com/federate")
 			if err != nil {
@@ -262,7 +300,7 @@ func TestQuery(t *testing.T) {
 					),
 				)
 				defer m.Close()
-				r := NewRoutes(m.url, proxyLabel)
+				r := NewRoutes(m.url, proxyLabel, false)
 
 				u, err := url.Parse("http://prometheus.example.com/api/v1/" + endpoint)
 				if err != nil {